+++ sentry/src/test/java/io/sentry/HubTest.kt
@@ -16,6 +16,9 @@ import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.TypeCheckHint.SENTRY_TYPE_CHECK_HINT
 import io.sentry.cache.EnvelopeCache
+import io.sentry.clientreport.ClientReportTestHelper.Companion.assertClientReport
+import io.sentry.clientreport.DiscardReason
+import io.sentry.clientreport.DiscardedEvent
 import io.sentry.hints.SessionEndHint
 import io.sentry.hints.SessionStartHint
 import io.sentry.protocol.SentryId
@@ -270,7 +273,7 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
     }
@@ -280,7 +283,7 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
         whenever(mockClient.captureEvent(any(), any(), anyOrNull())).thenReturn(SentryId(UUID.randomUUID()))
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         val lastEventId = sut.lastEventId
         sut.close()
@@ -293,7 +296,7 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
         verify(mockClient, never()).captureSession(any(), any())
@@ -304,7 +307,7 @@ class HubTest {
         val (sut, mockClient) = getEnabledHub()
 
         val event = SentryEvent()
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
         verify(mockClient, never()).captureSession(any(), any())
@@ -320,7 +323,7 @@ class HubTest {
 
         val event = SentryEvent(exception)
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         assertEquals(span.spanContext, event.contexts.trace)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
@@ -336,7 +339,7 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException(rootCause))
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         assertEquals(span.spanContext, event.contexts.trace)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
@@ -353,7 +356,7 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException(exceptionAssignedToSpan))
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         assertEquals(span.spanContext, event.contexts.trace)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
@@ -371,7 +374,7 @@ class HubTest {
         val originalSpanContext = SpanContext("op")
         event.contexts.trace = originalSpanContext
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         assertEquals(originalSpanContext, event.contexts.trace)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
@@ -383,7 +386,7 @@ class HubTest {
 
         val event = SentryEvent(RuntimeException())
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to { })
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to { })
         sut.captureEvent(event, hintsMap)
         assertNull(event.contexts.trace)
         verify(mockClient).captureEvent(eq(event), any(), eq(hintsMap))
@@ -452,7 +455,7 @@ class HubTest {
     fun `when captureException is called with a valid argument and hint, captureEvent on the client should be called`() {
         val (sut, mockClient) = getEnabledHub()
 
-        val hintsMap = mutableMapOf<String, Any?>(SENTRY_TYPE_CHECK_HINT to Object())
+        val hintsMap = mutableMapOf<String, Any>(SENTRY_TYPE_CHECK_HINT to Object())
         sut.captureException(Throwable(), hintsMap)
         verify(mockClient).captureEvent(any(), any(), any())
     }
@@ -1172,6 +1175,25 @@ class HubTest {
         sentryTracer.finish()
         val traceState = sentryTracer.traceState()
         verify(mockClient, never()).captureTransaction(any(), eq(traceState), any(), eq(null), anyOrNull())
+    }
+
+    @Test
+    fun `transactions lost due to sampling are recorded as lost`() {
+        val options = SentryOptions()
+        options.cacheDirPath = file.absolutePath
+        options.dsn = "https://key@sentry.io/proj"
+        options.setSerializer(mock())
+        val sut = Hub(options)
+        val mockClient = mock<ISentryClient>()
+        sut.bindClient(mockClient)
+
+        val sentryTracer = SentryTracer(TransactionContext("name", "op", false), sut)
+        sentryTracer.finish()
+
+        assertClientReport(
+            options.clientReportRecorder,
+            listOf(DiscardedEvent(DiscardReason.SAMPLE_RATE.reason, DataCategory.Transaction.category, 1))
+        )
     }
     //endregion